fix files

+++ C11/ex05/include/operator.h
@@ -6,7 +6,7 @@
 /*   By: sooykim <sooykim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/22 02:34:16 by sooykim           #+#    #+#             */
-/*   Updated: 2021/09/22 02:37:19 by sooykim          ###   ########.fr       */
+/*   Updated: 2021/09/25 01:31:20 by sooykim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@
 # define DIV 3
 # define MOD 4
 
-int	add(int a, int b);
-int	sub(int a, int b);
-int	mul(int a, int b);
-int	div(int a, int b);
-int	mod(int a, int b);
+long long int	add(int a, int b);
+long long int	sub(int a, int b);
+long long int	mul(int a, int b);
+long long int	div(int a, int b);
+long long int	mod(int a, int b);
 
 #endif
