+++ C11/ex05/src/parser.c
@@ -6,7 +6,7 @@
 /*   By: sooykim <sooykim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/22 02:38:07 by sooykim           #+#    #+#             */
-/*   Updated: 2021/09/25 00:59:11 by sooykim          ###   ########.fr       */
+/*   Updated: 2021/09/25 01:36:17 by sooykim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 int	parse_int(char *str)
 {
-	int	mcnt;
-	int	num;
+	int				mcnt;
+	long long int	num;
 
 	mcnt = 0;
 	num = 0;
